+++ projecteorigen/projectebase-api/src/main/java/es/caib/projectebase/api/services/UnitatOrganicaResource.java
@@ -5,6 +5,7 @@ import es.caib.projectebase.ejb.UnitatOrganicaService;
 import es.caib.projectebase.persistence.UnitatOrganica;
 import org.eclipse.microprofile.openapi.annotations.Operation;
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
+import org.eclipse.microprofile.openapi.annotations.headers.Header;
 import org.eclipse.microprofile.openapi.annotations.media.Content;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
@@ -70,7 +71,7 @@ public class UnitatOrganicaResource {
             description = "Unitat orgànica",
             content = @Content(mediaType = "application/json",
                     schema = @Schema(implementation = UnitatOrganica.class)))
-    @APIResponse(responseCode = "404", description = "Unitat orgànica no trobada")
+    @APIResponse(responseCode = "404", description = "Recurs no trobat")
     public Response get(@Parameter(description = "L'identificador de la unitat", required = true)
                         @PathParam("id") Long id) {
         UnitatOrganica unitatOrganica = unitatOrganicaService.findById(id);
@@ -89,7 +90,8 @@ public class UnitatOrganicaResource {
      */
     @POST
     @Operation(operationId = "createUnitat", summary = "Crea una nova unitat orgànica")
-    @APIResponse(responseCode = "201", description = "L'enllaç a la unitat orgànica creada")
+    @APIResponse(responseCode = "201", description = "Recurs creat",
+            headers = @Header(name = "location", description = "Enllaç al nou recurs"))
     public Response create(
             @RequestBody(
                     description = "Unitat orgànica",
@@ -111,8 +113,8 @@ public class UnitatOrganicaResource {
     @PUT
     @Path("{id}")
     @Operation(operationId = "updateUnitat", summary = "Actualitza una unitat orgànica")
-    @APIResponse(responseCode = "200", description = "La modificació s'ha realitzat correctament")
-    @APIResponse(responseCode = "404", description = "La unitat orgànica que es vol modificar no existeix.")
+    @APIResponse(responseCode = "204", description = "Operació realitzada correctament")
+    @APIResponse(responseCode = "404", description = "Recurs no trobat")
     public Response update(
             @RequestBody(
                     description = "Unitat orgànica",
@@ -131,7 +133,7 @@ public class UnitatOrganicaResource {
             unitatActual.setEstat(unitatOrganica.getEstat());
             unitatActual.setDataCreacio(unitatOrganica.getDataCreacio());
             unitatOrganicaService.update(unitatActual);
-            return Response.ok().build();
+            return Response.noContent().build();
         }
     }
 
@@ -144,11 +146,15 @@ public class UnitatOrganicaResource {
     @DELETE
     @Path("{id}")
     @Operation(operationId = "deleteUnitat", summary = "Esborra una unitat orgànica")
-    @APIResponse(responseCode = "200", description = "La unitat s'ha esborrat correctament")
-    @APIResponse(responseCode = "404", description = "Unitat orgànica no trobada")
+    @APIResponse(responseCode = "204", description = "Operació realitzada correctament")
+    @APIResponse(responseCode = "404", description = "Recurs no trobat")
     public Response delete(@Parameter(description = "L'identificador de la unitat", required = true)
                         @PathParam("id") Long id) throws I18NException {
-        unitatOrganicaService.deleteById(id);
-        return Response.ok().build();
+        if (unitatOrganicaService.findById(id) == null) {
+            return Response.status(Response.Status.NOT_FOUND).build();
+        } else {
+            unitatOrganicaService.deleteById(id);
+            return Response.noContent().build();
+        }
     }
 }
